Cleaned up the API a little v2
