--- stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/DefaultMessageListEmptyContent.kt
@@ -1,32 +1,0 @@
-package io.getstream.chat.android.compose.ui.components.messages
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import io.getstream.chat.android.compose.R
-import io.getstream.chat.android.compose.ui.theme.ChatTheme
-
-/**
- * The default empty placeholder that is displayed when there are no messages in the channel.
- *
- * @param modifier Modifier for styling.
- */
-@Composable
-internal fun DefaultMessageListEmptyContent(modifier: Modifier) {
-    Box(
-        modifier = modifier.background(color = ChatTheme.colors.appBackground),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            text = stringResource(R.string.stream_compose_message_list_empty_messages),
-            style = ChatTheme.typography.body,
-            color = ChatTheme.colors.textLowEmphasis,
-            textAlign = TextAlign.Center
-        )
-    }
-}
