+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/list/MessageList.kt
@@ -1,15 +1,22 @@
 package io.getstream.chat.android.compose.ui.messages.list
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.state.imagepreview.ImagePreviewResult
 import io.getstream.chat.android.compose.state.imagepreview.ImagePreviewResultType
 import io.getstream.chat.android.compose.state.messages.MessagesState
 import io.getstream.chat.android.compose.state.messages.list.GiphyAction
 import io.getstream.chat.android.compose.state.messages.list.MessageListItemState
 import io.getstream.chat.android.compose.ui.components.LoadingIndicator
-import io.getstream.chat.android.compose.ui.components.messages.DefaultMessageListEmptyContent
+import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.viewmodel.messages.MessageListViewModel
 
 /**
@@ -107,6 +114,26 @@ internal fun DefaultMessageContainer(
 @Composable
 internal fun DefaultMessageListLoadingIndicator(modifier: Modifier) {
     LoadingIndicator(modifier)
+}
+
+/**
+ * The default empty placeholder that is displayed when there are no messages in the channel.
+ *
+ * @param modifier Modifier for styling.
+ */
+@Composable
+internal fun DefaultMessageListEmptyContent(modifier: Modifier) {
+    Box(
+        modifier = modifier.background(color = ChatTheme.colors.appBackground),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = stringResource(R.string.stream_compose_message_list_empty_messages),
+            style = ChatTheme.typography.body,
+            color = ChatTheme.colors.textLowEmphasis,
+            textAlign = TextAlign.Center
+        )
+    }
 }
 
 /**
